Generate unique IDs for each sketch
We can't use the count directly, because then all sketches (including
e.g. sketches in workers) would generate the same sequence. That would
result in duplicate values, and thus various issues in estimation.

So each sketch gets a unique seed, and then combines that with the count
using XXH32(). Maybe it'd be enough to use a shorter seed and just
append the count (without the hash), but that seems like it would be
more likely to generate duplicate values. And we may want to only use
a shorter ID, so we'd need a hash to mix the bits anyway.

There's a bit more work to be done to handle duplicate IDs (even if not
very unlikely). Right now it just falls over (in assert build), and
might cause some misbehavior in regular builds. It's probably enough to
just remove the duplicate values (as it should be very rare), or maybe
we could replace it with a random value (but that may take time to get
a low enough hash).

The unfortunate consequence is that this makes the results a bit random.
This matters for regression testing - the tests can't expect expect
exact numbers, for example. Instead, the tests check for a range. It's
not clear if this is a good long-time solution, it might be flaky.